Remove sublime and eclipse
Sublime is not as popular as before and causes the duplication of
runtimes due to its outdated dependencies. The removal of this IDE will
remove the extra copies of `org.freedesktop.Platform.GL.default`,
`org.freedesktop.Sdk.Locale` and `org.freedesktop.Sdk`.

Eclipse cannot be installed using the packaged Flatpak on Ubuntu 20.04,
so it had to be removed even though it is still a popular IDE.

- Remove Sublime (Flatpak)
- Remove Eclipse IDE (Flatpak)