Fix required download for C/C++ extension
The C and C++ extension downloaded from the online repository required
the installation of downloadable assets in order to work correctly. This
situation would have lead to a broken development environment (only
compilation and execution would be functional while suggestions and
debugging would be broken) while the competition was in progress, as the
internet is disabled by a firewall while the contest is being held.

This commit downloads the fully featured offline version of the
`ms-vscode.cpptools` instead of the standard one available in the
extensions store. The newest release will be downloaded from the
official repository using the Github API, which required the addition of
the packages `jq` and `wget`. No integrity checks are performed as the
official Microsoft repository is believed to be *trustworthy* and no
signature or hash files are available in the release assets offered.

- Adds `jq` and `wget` in the `Depends` section of `debian/control`
- Downloads and installs offline version of `ms-vscode.cpptools` from
  the official repository using Github's API.