Update maintainer, sections and dependencies
The maintainer name has been updated to fit with the other packages.

The packages `maratona-kotlin-doc` and `maratona-vscode-extensions`
contained a blank line as the beginning of their extended description.
This is not recommended, and it triggered a Lintian warning
`extended-description-contains-empty-paragraph`. In order to fix this
problem, the descriptions have been updated. Also, `maratona-kotlin-doc`
is not considered a safe package any more.

Packages `maratona-linguagens-doc` and `maratona-kotlin-doc` did not
have their sections specified as recommended. In order to fix this
problem pointed by Lintian (`wrong-section-according-to-package-name`),
those packages were moved to the Documentation (`doc`) section.

Finally, the packages `maratona-linguagens` and
`maratona-editores-flatpak` used `debconf` without specifying it as a
dependency. Lintian flagged this as a warning, as `debconf` is not an
essential package (`missing-debconf-dependency`). This problem was fixed
adding `debconf` to the `Depends` field.

The packages `maratona-editores-flatpak`, `maratona-editores` and
`maratona-linguagens-doc` had their descriptions updated. This fixed
broken lines and added details about which resources will be added to
the system.

The package `maratona-vscode-extensions` had its description shortened
to fit in the 80 characters length. This solved the Lintian warning
`synopsis-too-long`. Also, the packages `maratona-kotlin-doc` and
`maratona-vscode-extensions` had to have their incorrectly formatted
description fields fixed in order to silence the
`synopsis-is-a-sentence` warning.

As requested by the maintainer, `valgrind` has been added as a
dependency to `maratona-linguagens` for helping with C/C++ debugging.
Information about the main repository on Github has also been added on
the control file. Also, the dependency `python3-distutils` has been
moved from `maratona-editores` to `maratona-linguagens`, keeping all
Python 3 language packages as dependencies of the latter.

The lintian version used was v2.104.0 (Debian Bullseye). The options
used were `-EviIL +pedantic --profile ubuntu`. The relevant issue in the
repository, issue #31 , was updated with the changes in this commit.

- updates `maratona-kotlin-doc` and `maratona-vscode-extensions`
  description's, fixing `extended-description-contains-empty-paragraph`
  and removing the statement that considered the former safe to install
  in any Debian or Ubuntu system;
- specifies the field `Section` as `doc` in `maratona-linguagens-doc`
  and `maratona-kotlin-doc`, fixing
  `wrong-section-according-to-package-name`;
- adds `debconf` to the `Depends` field of `maratona-linguagens` and
  `maratona-editores-flatpak`, fixing `missing-debconf-dependency`;
- Adds `valgrind` as a dependency of `maratona-linguagens`;
- Shortens `maratona-vscode-extensions`'s description to 80 chars
  (solves `synopsis-too-long`);
- Fixes formatting of `maratona-vscode-extensions` and
  `maratona-kotlin-doc` descriptions, fixing `synopsis-is-a-sentence`.
- Updates the descriptions of `maratona-editores-flatpak`,
  `maratona-editores` and `maratona-linguagens-doc`;
- Move `python3-distutils` from `maratona-editores` to
  `maratona-linguagens`;
- adds Vcs-Browser and Vcs-Git fields on `debian/control`