Avoid running VSCode as root and fix extensions
The extensions will not be downloaded with root privileges using the
post installation script. Recent changes in the Visual Studio Flatpak
rendered impossible to run the package as root, which was already
inadequate and frowned upon by the community as Flatpak will eve print
a warning if it is run by sudo.

The solution found is to prepare the extensions in the development
environment and ship them pre-installed in the binary Debian package
responsible for installing them. This alone will not install the
extensions, as they need to be transferred to a directory with read and
write permissions and Flatpak does not reliably mount files stored in
the /var directory.

The complete installation will only happen with the modifications made
in the zera-home-icpc script installed by maratona-usuario-icpc or any
of the packages that depend on it. The pre-installed extensions must be
copied into the icpc's home directory, where Visual Studio Code will
perform the necessary modifications in order to update the files during
the usage.

The Java extension was removed as it keeps failing to find the system
JDK. IntelliJ, also installed as a Flatpak, has no trouble doing it and
supports both Java and Kotlin without any difficulties.

The Python extension has troubles finding the optional linters, but
still works as expected.

- remove post installation script
- ship formulahendry.code-runner v0.11.5, ms-python.python
  v2021.9.1246542782 and vscode-cpptools v1.6.0 in a pre-installed
  state, extracted from their original VSIX packages
- modify the rules so that dh will not mess with the pre-compiled stuff
  (binaries and libraries) used by the extensions
- remove the dependency on jq and wget, as downloading and parsing
  Github's JSON API will not be necessary anymore
- change maratona-vscode-extensions Architecture to amd64, as the C/C++
  extension by Microsoft is architecture dependent
- the pre-dependency on maratona-editores-flatpak is now unnecessary, so
  it has been changed into a regular dependency, as there is no
  necessity to have the editor fully configured before moving the
  extensions to their installation directory
- change the installation directory from /usr/share to /var/lib, as the
  files are not architecture independent and may be subjected to
  modifications