Depend on flatpaked OpenJDK 11
The flatpak packages are not currently capable of using Ubuntu's 22.04
packaged OpenJDKs due to an older version of Glibc included on the jails
that depend on Freedesktop 21.08 SDK.

This commit is basically a hotfix. The addition of flatpak's OpenJDK 11
may bloat the system, yet it will keep everything working until the
release of Freedesktop SDK 22.08.

- depend on flathub's OpenJDK until Freedesktop SDK 22.08 upgrades the
  jail's glibc